requirements.txt
Django==3.1.6
djangorestframework==3.11.0
Markdown==3.3.3
